Finish PL refactor
